Attempt to make the Game more scaleable.
Register the processes globally so that it can be run on multiple nodes.
Spawn a game server per user, and synchronize individual state to a global state backedup to ets,
so that the game server will not be the bottleneck, as its mailbox gets filled with messages.
The bottleneck will now shift to each player's genserver.

TODO: Explore any edgecase with consitancy and race-conditions in tests
Signed-off-by: Manu S Ajith <neo@codingarena.in>